refactor: Extracted functions (removed duplicate code)
