+++ ember-codemod-remove-ember-css-modules/src/migration/ember-app/steps/shared/index.js
@@ -1,0 +1,1 @@
+export * from './import-styles.js';
